[enhancement] Updated callback name to more understandable function names
auth_token_request means that a request is sent to someone to get back a secret (pin, user pin, pet name confirmation)
auth_token_acknowledge means that the smartcard confirmation is send back. The check has been done bu the lower layer and this is the (positive or negative) acknowledge.